Move unit checks to inp_functions
